Adds Players

+++ Controllers/GameNightsController.cs
@@ -82,6 +82,11 @@ namespace GameNightWithFriends.Controllers
                 return BadRequest();
             }
 
+            if (gameNight.MinimumPlayers < 2)
+            {
+                return BadRequest(new { Message = "You need at least 2 players! " });
+            }
+
             // Tell the database to consider everything in gameNight to be _updated_ values. When
             // the save happens the database will _replace_ the values in the database with the ones from gameNight
             _context.Entry(gameNight).State = EntityState.Modified;
@@ -125,6 +130,10 @@ namespace GameNightWithFriends.Controllers
         [HttpPost]
         public async Task<ActionResult<GameNight>> PostGameNight(GameNight gameNight)
         {
+            if (gameNight.MinimumPlayers < 2)
+            {
+                return BadRequest(new { Message = "You need at least 2 players! " });
+            }
             // Indicate to the database context we want to add this new record
             _context.GameNights.Add(gameNight);
             await _context.SaveChangesAsync();
@@ -150,7 +159,10 @@ namespace GameNightWithFriends.Controllers
                 // There wasn't a gameNight with that id so return a `404` not found
                 return NotFound();
             }
-
+            if (gameNight.When < DateTime.Now)
+            {
+                return BadRequest();
+            }
             // Tell the database we want to remove this record
             _context.GameNights.Remove(gameNight);
